team_management/requirements.txt
Django==1.9
django-localflavor==1.6.1
djangorestframework==3.7.3
drf-tracking==1.3.1
mysqlclient==1.3.12
pytz==2017.3
